Fix compilation error and warning with some tests
- Add the use of enable_if to not call update_buffer_task in the buffer
  destructor if the buffer is a read-only buffer with a const type

- Fix a warning caused by a comparison between signed and unsigned values

+++ tests/buffer/buffer_data_tracking.cpp
@@ -70,7 +70,7 @@ auto test2 = [](auto program, auto b_queue) {
 
   auto acc1 = A.get_access<access::mode::discard_write>();
 
-  for (int i = 0; i < N; ++i) acc1[i] += 1;
+  for (size_t i = 0; i < N; ++i) acc1[i] += 1;
 
   VERIFY_COND(A.is_data_up_to_date(host_context));
   VERIFY_COND(!A.is_data_up_to_date(device_context));
